# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Bank Server",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Ruhan CS",
            "url": "ruhancorreasoares@gmail.com",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/account": {
            "post": {
                "description": "create account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Create account",
                "parameters": [
                    {
                        "description": "account request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto_account.InputCreateAccountDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto_account.OutputCreateAccountDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.JsonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.JsonResponse"
                        }
                    }
                }
            }
        },
        "/account/credit": {
            "post": {
                "description": "Credit value on account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Credit value",
                "parameters": [
                    {
                        "description": "account request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto_account.InputCreditValueUseCase"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto_account.OutputCreditValueUseCase"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.JsonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.JsonResponse"
                        }
                    }
                }
            }
        },
        "/account/debit": {
            "post": {
                "description": "Credit value on account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Credit value",
                "parameters": [
                    {
                        "description": "account request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto_account.InputDebitValueUseCase"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto_account.OutputDebitValueUseCase"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.JsonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.JsonResponse"
                        }
                    }
                }
            }
        },
        "/account/transfer": {
            "post": {
                "description": "Credit value on account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Credit value",
                "parameters": [
                    {
                        "description": "account request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto_account.InputTransferUseCase"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto_account.OutputTransferUseCase"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.JsonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.JsonResponse"
                        }
                    }
                }
            }
        },
        "/user": {
            "post": {
                "description": "create user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create user",
                "parameters": [
                    {
                        "description": "user request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user_dto.InputCreateUserDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/user_dto.OutputCreateUserDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.JsonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.JsonResponse"
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "description": "Get user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user_dto.OutputGetUserDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
        "dto_account.InputCreateAccountDto": {
            "type": "object",
            "properties": {
                "owner": {
                    "type": "string"
                }
            }
        },
        "dto_account.InputCreditValueUseCase": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
        "dto_account.InputDebitValueUseCase": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
        "dto_account.InputTransferUseCase": {
            "type": "object",
            "properties": {
                "from_account": {
                    "type": "string"
                },
                "to_account": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
        "dto_account.OutputCreateAccountDto": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                }
            }
        },
        "dto_account.OutputCreditValueUseCase": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
        "dto_account.OutputDebitValueUseCase": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
        "dto_account.OutputTransferUseCase": {
            "type": "object",
            "properties": {
                "from_account": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "to_account": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
        "user_dto.InputCreateUserDto": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "user_dto.OutputCreateUserDto": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "username": {
                    "description": "ID       string `json:\"id\"`",
                    "type": "string"
                }
            }
        },
        "user_dto.OutputGetUserDto": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "web.JsonResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}